requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
protobuf==4.21.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyTelegramBotAPI==4.6.1
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
uritemplate==4.1.1
urllib3==1.26.11
